services/whisper-transcription-service/requirements.txt
# Whisper Transcription Service Requirements
# High-accuracy Italian voice transcription using OpenAI Whisper

# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0

# Whisper and audio processing
faster-whisper==1.0.0
torch==2.1.0
torchaudio==2.1.0
numpy==1.24.3
scipy==1.11.4
librosa==0.10.2.post1
soundfile==0.13.1

# NVIDIA CUDA libraries for GPU support
nvidia-cublas-cu11==11.10.3.66
nvidia-cudnn-cu11==8.9.2.26

# File handling and utilities
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
aiofiles==23.2.1

# HTTP client
httpx==0.25.2

# Logging and monitoring
structlog==23.2.0

# Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
webrtcvad==2.0.10
